+++ src/apply.jl
@@ -1,132 +1,123 @@
-MATH_DOTONLY    = [:.+, :.-, :.*, :./, :.^]
-MATH_ALL        = [MATH_DOTONLY; [:+, :-, :*, :/, :^]]
-COMPARE_DOTONLY = [:.>, :.<, :.==, :.>=, :.<=] 
+UNARY = [:+, :-, :~, :!]
+MATH_DOTONLY    = [:.+, :.-, :.*, :./, :.%, :.^]
+MATH_ALL        = [MATH_DOTONLY; [:+, :-, :*, :/, :%]]
+COMPARE_DOTONLY = [:.>, :.<, :.==, :.>=, :.<=, :.!=]
+BOOLEAN_OPS     = [:&; :|; :$; COMPARE_DOTONLY]
 
-for op in [MATH_ALL; COMPARE_DOTONLY]
+for op in [UNARY; MATH_ALL; BOOLEAN_OPS]
     eval(Expr(:import, :Base, op))
 end # for
 
-###### Mathematical operators  ###############
+###### Unary operators and functions ################
 
-# TimeArray <--> TimeArray 
-for op in MATH_DOTONLY
+# TimeArray
+for op in UNARY
     @eval begin
-        function ($op){T,N}(ta1::TimeArray{T,N}, ta2::TimeArray{T,N})
-          # first test metadata matches
-          ta1.meta == ta2.meta ? meta = ta1.meta : error("metadata doesn't match")
-          cname  = [ta1.colnames[1][1:2] *  string($op) *  ta2.colnames[1][1:2]]
-          idx1, idx2 = overlaps(ta1.timestamp, ta2.timestamp)
-          # obtain shared timestamp
-          tstamp = ta1[idx1].timestamp
-          # retrieve values that match the Int array matching dates
-          vals1  = ta1[idx1].values
-          vals2  = ta2[idx2].values
-          vals = Array(T, length(idx1))
-          for i in 1:length(vals)
-              vals[i] = ($op)(vals1[i], vals2[i])
-          end
-          TimeArray(tstamp, vals, cname, meta)
+        function ($op){T,N}(ta::TimeArray{T,N})
+            cnames  = [string($op) * name for name in ta.colnames]
+            vals = ($op)(ta.values)
+            TimeArray(ta.timestamp, vals, cnames, ta.meta)
         end # function
     end # eval
 end # loop
 
-# TimeArray (2d) <--> TimeArray (1d)
-for op in MATH_DOTONLY
+###### Numerical operations and comparisons #########
+
+# TimeArray <--> Scalar
+for op in [MATH_ALL; COMPARE_DOTONLY]
     @eval begin
-        function ($op){T}(ta1::TimeArray{T,2}, ta2::TimeArray{T,1})
-           # first test metadata matches
-           ta1.meta == ta2.meta ? meta = ta1.meta : error("metadata doesn't match")
-           # interate to find when there is a match on timestamp
-           counter = Int[]
-           for i in 1:length(ta1)
-               if in(ta1[i].timestamp[1], ta2.timestamp)
-                   push!(counter, i)
-               end
-           end
-           # create new shortened versions of ta1 and ta2
-           newta1 = ta1[counter]
-           newta2 = ta2[newta1.timestamp]
-
-           # operate on the values columns
-           vals = ($op)(ta1.values, ta2.values) 
-
-           cnames = repmat([""], length(ta1.colnames))
-           for i in 1:length(ta1.colnames)
-               cnames[i] = string(ta1.colnames[i])[1:2] * " " *  string($op) * " " *  string(ta2.colnames[1])[1:2]
-           end
-           TimeArray(newta1.timestamp, vals, cnames, meta)
+        function ($op){T<:Number,N}(ta::TimeArray{T,N}, var::Number)
+            cnames  = [name * string($op) * string(var) for name in ta.colnames]
+            vals = ($op)(ta.values, var)
+            TimeArray(ta.timestamp, vals, cnames, ta.meta)
         end # function
     end # eval
 end # loop
 
-# TimeArray <--> Int,Float64
-for op in MATH_ALL
+# Scalar <--> TimeArray
+for op in [MATH_ALL; COMPARE_DOTONLY]
     @eval begin
-        function ($op){T,N}(ta::TimeArray{T,N}, var::Union(Int,Float64))
-            vals = ($op)([t for t in ta.values], var)
-            TimeArray(ta.timestamp, vals, ta.colnames, ta.meta)
+        function ($op){T<:Number,N}(var::Number, ta::TimeArray{T,N})
+            cnames  = [string(var) * string($op) * name for name in ta.colnames]
+            vals = ($op)(var, ta.values)
+            TimeArray(ta.timestamp, vals, cnames, ta.meta)
         end # function
     end # eval
 end # loop
 
-# element-wise mathematical operations between an Int,Float64 and column
-for op in MATH_ALL
+# ND TimeArray <--> MD TimeArray
+for op in [MATH_DOTONLY; COMPARE_DOTONLY]
     @eval begin
-        function ($op){T,N}(var::Union(Int,Float64), ta::TimeArray{T,N})
-            vals = ($op)(var, [t for t in ta.values])
-            TimeArray(ta.timestamp, vals, ta.colnames, ta.meta)
+        function ($op){T<:Number,N,M}(ta1::TimeArray{T,N}, ta2::TimeArray{T,M})
+            # first test metadata matches
+            ta1.meta == ta2.meta ? meta = ta1.meta : error("metadata doesn't match")
+            # determine array widths and name cols accordingly
+            w1, w2  = length(ta1.colnames), length(ta2.colnames)
+            if w1 == w2
+              cnames = [ta1.colnames[i]*string($op)*ta2.colnames[i] for i=1:w1]
+            elseif w1==1
+              cnames = [ta1.colnames[1]*string($op)*ta2.colnames[i] for i=1:w2]
+            elseif w2==1
+              cnames = [ta1.colnames[i]*string($op)*ta2.colnames[1] for i=1:w1]
+            else
+              error("arrays must have the same number of columns, or one must be a single column")
+            end
+            # obtain shared timestamp
+            idx1, idx2 = overlaps(ta1.timestamp, ta2.timestamp)
+            tstamp = ta1[idx1].timestamp
+            # retrieve values that match the Int array matching dates
+            vals1, vals2 = ta1[idx1].values, ta2[idx2].values
+            # compute output values
+            vals = ($op)(vals1, vals2)
+            TimeArray(tstamp, vals, cnames, meta)
         end # function
     end # eval
 end # loop
 
-###### Comparison operators  ###############
+###### Boolean operations and comparisons ###############
 
-# TimeArray <--> TimeArray 
-for op in COMPARE_DOTONLY 
+# TimeArray <--> Bool
+for op in BOOLEAN_OPS
     @eval begin
-        function ($op){T,N}(ta1::TimeArray{T,N}, ta2::TimeArray{T,N})
-          # first test metadata matches
-          ta1.meta == ta2.meta ? meta = ta1.meta : error("metadata doesn't match")
-          cname  = [ta1.colnames[1][1:2] *  string($op) *  ta2.colnames[1][1:2]]
-          idx1, idx2 = overlaps(ta1.timestamp, ta2.timestamp)
-          # obtain shared timestamp
-          tstamp = ta1[idx1].timestamp
-          # retrieve values that match the Int array matching dates
-          vals1  = ta1[idx1].values
-          vals2  = ta2[idx2].values
-          vals = Array(Bool, length(idx1))
-          for i in 1:length(vals)
-              vals[i] = ($op)(vals1[i], vals2[i])
-          end
-          TimeArray(tstamp, vals, cname, meta)
+        function ($op){N}(ta::TimeArray{Bool,N}, var::Bool)
+            cnames = [name * string($op) * string(var) for name in ta.colnames]
+            vals   = ($op)(ta.values, var)
+            TimeArray(ta.timestamp, vals, cnames, ta.meta)
         end # function
     end # eval
 end # loop
 
-# TimeArray <--> Int,Float64
-for op in COMPARE_DOTONLY 
+# Bool <--> TimeArray
+for op in BOOLEAN_OPS
     @eval begin
-        function ($op){T,N}(ta::TimeArray{T,N}, var::Union(Int,Float64))
-          cname  = [ta.colnames[1][1:2] *  string($op) *  string(var)]
-          vals   = Array(Bool, length(ta))
-          for i in 1:length(vals)
-              vals[i] = ($op)(ta.values[i], var) 
-          end
-          TimeArray(ta.timestamp, vals, cname, ta.meta)
+        function ($op){N}(var::Bool, ta::TimeArray{Bool,N})
+            cnames = [string(var) * string($op) * name for name in ta.colnames]
+            vals   = ($op)(var, ta.values)
+            TimeArray(ta.timestamp, vals, cnames, ta.meta)
         end # function
     end # eval
 end # loop
 
-# Int,Float64 <--> TimeArray
-for op in COMPARE_DOTONLY 
+# Boolean ND TimeArray <--> Boolean ND TimeArray
+# Doesn't support broadcasting since it isn't supported by Base either
+for op in BOOLEAN_OPS
     @eval begin
-        function ($op){T,N}(var::Union(Int,Float64), ta::TimeArray{T,N})
-          cname  = [ta.colnames[1][1:2] *  string($op) *  string(var)]
-          vals   = Array(Bool, length(ta))
-          for i in 1:length(vals)
-              vals[i] = ($op)(var, ta.values[i])
-          end
-          TimeArray(ta.timestamp, vals, cname, ta.meta)
+        function ($op){N}(ta1::TimeArray{Bool,N}, ta2::TimeArray{Bool,N})
+            # test column count matches
+            length(ta1.colnames) == length(ta2.colnames) ?
+              ncols=length(ta1.colnames) :
+              error("arrays must have the same number of columns")
+            # test metadata matches
+            ta1.meta == ta2.meta ? meta = ta1.meta : error("metadata doesn't match")
+            cnames = [ta1.colnames[i]*string($op)*ta2.colnames[i] for i=1:ncols]
+            idx1, idx2 = overlaps(ta1.timestamp, ta2.timestamp)
+            # obtain shared timestamp
+            tstamp = ta1[idx1].timestamp
+            # retrieve values that match the Int array matching dates
+            vals1  = ta1[idx1].values
+            vals2  = ta2[idx2].values
+            vals = ($op)(vals1, vals2)
+            TimeArray(tstamp, vals, cnames, meta)
         end # function
     end # eval
 end # loop